+++ server/src/main/java/today/wtfood/server/service/EventService.java
@@ -39,14 +39,14 @@ public class EventService {
     }
 
     // ID로 이벤트 조회
-    public EventDetail getEventById(int id) {
+    public EventDetail getEventById(long id) {
         return er.findDetailById(id)
                 .orElseThrow(() -> new RuntimeException("Event with id " + id + " not found"));
     }
 
     // 이벤트 수정
     @Transactional
-    public void updateEvent(int id, EventDto dto) {
+    public void updateEvent(long id, EventDto dto) {
         Event event = er.findById(id)
                 .orElseThrow(() -> new RuntimeException("Event with id " + id + " not found"));
         event.setTitle(dto.getTitle());
@@ -57,7 +57,7 @@ public class EventService {
     }
 
     // 이벤트 삭제
-    public void deleteEvent(int id) {
+    public void deleteEvent(long id) {
         if (er.existsById(id)) {
             er.deleteById(id);
         } else {
